+++ CHANGELOG.rst
@@ -7,12 +7,16 @@ Added
 - Added support for Internationalized Domain Names (IDN) in :meth:`~dhcpkit.utils.parse_domain_bytes` and
   :meth:`~dhcpkit.utils.encode_domain`. This makes it possible to use IDN everywhere in DHCPKit, including configuration
   files.
+- Implement a domain socket to control the server process.
+- Added `ipv6-dhcpctl` to control the server process through the domain socket.
 
 Changed
 ^^^^^^^
 
 - Created ForOtherServerError as a subclass of CannotRespondError, to enable more accurate logging, and to make it
   possible to gather better statistics.
+- Create PID file /var/run/ipv6-dhcpd.pid by default.
+- Create domain socket /var/run/ipv6-dhcpd.sock control the server by default.
 
 Deprecated
 ^^^^^^^^^^